+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## [0.2.0] - December 02, 2022
+* sdk: ">=2.18.0 <3.0.0"
+* Removed underscore from variables in the private State class.
+* Supplied all the Draggable widget's callback functions.
+* Change the required parameter from Widget to FloatingActionButton
+
 ## [0.1.4] - October 18, 2021
 * Fixed issue when running using dart 2.12  and above